+++ src/util/index.js
@@ -1,1 +1,2 @@
 export * as default from './util';
+export * as schedule from './schedule';